fix: switch gwei command source
