+++ gamestonk_terminal/cryptocurrency/onchain/onchain_controller.py
@@ -15,7 +15,10 @@ from gamestonk_terminal.helper_funcs import (
     check_int_range,
 )
 
-from gamestonk_terminal.cryptocurrency.onchain import gasnow_view, whale_alert_view
+from gamestonk_terminal.cryptocurrency.onchain import (
+    ethgasstation_view,
+    whale_alert_view,
+)
 
 # pylint: disable=R1732
 
@@ -102,7 +105,7 @@ class OnchainController:
             prog="onchain",
             description="""
                 Display ETH gas fees
-                [Source: https://www.gasnow.org]
+                [Source: https://ethgasstation.info]
             """,
         )
 
@@ -121,7 +124,7 @@ class OnchainController:
             if not ns_parser:
                 return
 
-            gasnow_view.display_gwei_fees(export=ns_parser.export)
+            ethgasstation_view.display_gwei_fees(export=ns_parser.export)
 
         except Exception as e:
             print(e, "\n")
